Dry up downloading logic
